WIP: working on converting the PostTools
