+++ src/components/assets/Icons.css
@@ -77,10 +77,6 @@
   transition-delay: var(--speed2), var(--speed), var(--speed);
 }
 
-.MarkerIcon {
-  color: #ababab;
-}
-
 /* ------------------------------------- */
 
 .ElloMark,
@@ -135,6 +131,63 @@
 .TabListStreamContainer .ElloMark.isSpinner > .SmileShape,
 .MutedUsers .ElloMark.isSpinner > .SmileShape {
   fill: #fff !important;
+}
+
+/* ------------------------------------- */
+
+/* stylelint-disable selector-max-compound-selectors, selector-max-specificity, selector-no-qualifying-type */
+.WatchTool > button.isActive > .BoltIcon > g,
+.Post .isCommentsActive .BubbleIcon > g,
+.PostDetails .Post .BubbleIcon > g,
+.Post .BubbleIcon > g:nth-child(2) {
+  fill: #000;
+  stroke: #000;
+}
+
+.no-touch .CommentTool > button:hover > .BubbleIcon > g,
+.no-touch .FlagTool > button:hover > .FlagIcon > g:nth-of-type(2) {
+  fill: #000;
+}
+
+.Post .isCommentsRequesting .BubbleIcon > g:nth-child(1),
+.Post .BubbleIcon > g:nth-child(2),
+.StyleGuide .BubbleIcon > g:nth-child(2) {
+  display: none;
+}
+
+.Post .isCommentsRequesting .BubbleIcon > g:nth-child(2) {
+  display: inline;
+}
+
+.Post .isCommentsRequesting .BubbleIcon polygon {
+  animation: animateCommentsLoading 0.75s infinite var(--ease-in-out-quart);
+}
+
+.PostTool .EyeIcon,
+.PostTool .BubbleIcon {
+  margin-top: -2px;
+}
+
+/* Animations ?? */
+.ViewsTool .EyeIcon > circle {
+  fill: currentColor;
+  stroke: transparent;
+  transition: fill var(--speed) ease, stroke var(--speed) ease;
+}
+
+.no-touch .ViewsTool > a:hover > .EyeIcon > circle {
+  stroke: #000;
+  transition-delay: var(--speed), 0s;
+}
+
+.no-touch .ViewsTool > a:hover > .EyeIcon > g {
+  stroke: #000;
+}
+
+/* stylelint-enable selector-max-compound-selectors, selector-max-specificity, selector-no-qualifying-type */
+
+.MarkerIcon {
+  color: #ababab;
 }
 
 .XClose {
