+++ src/components/omnibar/Omnibar.css
@@ -40,7 +40,7 @@
 
 .Omnibar .ChevronIcon {
   margin-right: 5px;
-  transform: rotate(180deg);
+  transform: rotate(-90deg);
 }
 
 /* // ------------------------------------- */
