+++ src/components/posts/PostTools.css
@@ -92,8 +92,6 @@
   color: #aaa;
 }
 
-.PostTool .EyeIcon,
-.PostTool .BubbleIcon,
 .PostTool .HeartIcon {
   margin-top: -2px;
 }
