+++ src/components/posts/PostTools.js
@@ -9,6 +9,13 @@ import {
   BubbleIcon,
   EyeIcon,
   FlagIcon,
+  // HeartIcon,
+  // PencilIcon,
+  // RepostIcon,
+  // ShareIcon,
+  // XBoxIcon,
+} from '../assets/Icons'
+import {
   HeartIcon,
   PencilIcon,
   RepostIcon,
